Enhance generator to work with x-y-z geometry files
For templates which have x-y-z arrays for geometries, we need to loop over the
arrays and manually insert them into the final xml as the simple parsing trick won't accept
complex tags which are not part of a sub tag.
This fix creates a higher level tag with all 3 inside, then strips away the top level
after the xdmf build() has been called and all structures created.