Add support for Dimensions="@Density@Dimensions" syntax
The purpose of this is to allow certain entries to specify values
which must be obtained from datasets. In this example we wish to obtain
the dimensions of the dataset from one of the scalar arrays.

The syntax is subject to change as it evolves to handle a wider variety of
cases.